chore: enrich and improve telemetry events  (#29)
- added `auth_token` to events after `auth_token_configured` (read it
from .localstack/auth.json file when possible)
- retry info endpoint [1] to improve `session_id` capture rate for
`emulator.started` events
- removed `expected_steps` from `setup_started` event
- renamed `position` to `step_order`
- changed `SCHEMA_VERSION`: 1 -> 2
- added `license_setup_ended` event
- changed steps for "setup_ended" event
- capture "setup_ended" when user cancels
- capture "skipped" events better.

------
[1] We were not capturing the emulator session_id quite well in the
emulator.started events.
We rely on endpoint http://localhost:4566/_localstack/info to grab it,
however the endpoint is not immediately available when we just start
localstack (needs to be in running state). It needs a couple of seconds
before it's accessible (some 4-6 seconds in my local tests).
Now we retry up to 10x and wait 1s each time before sending the event to
make sure we get the session_id as often as possible.