Show Setup Wizard pop-up once setup is required
This ensures that pop-up appears if setup is required at the start. Instead of synchronizing changes at the start we decided to have pop-up appearing every time the setup is broken. This shouldn't happen normally, but when it happens then setup wizard pop-up in addition to painting setup bar red can actually be helpful.