fix: show LocalStack setup on only when setup is required (#58)
We've had a couple of race conditions and missing status check
dependencies. This resulted in LocalStack Setup notification being shown
when the setup is not required - this was happening on the most of VS
Code starts.

This PR fixes it with the following:

- Add explicit status initialization on every status tracker creation.
- Show Setup Wizard pop-up once setup is required. This ensures that
pop-up appears if setup is required at the start. Instead of
synchronizing statuses at the start and ensuring Setup Wizard is only
popping up once, we decided to have pop-up appearing every time the
setup is broken. This shouldn't happen normally, but when it happens
then setup wizard pop-up in addition to painting setup bar red can
actually be helpful.
- License status tracker now watches both license and auth token file
(not very clean solution, in the next iteration we should add behavioral
dependency between license tracker and auth tracker).
- Refresh status bar is now happening after the last setup wizard step.
Before it was in the middle.